GovInfo/Gov/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
beautifulsoup4==4.12.3      
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2   
Django==5.0.3
djangorestframework==3.15.1 
h11==0.14.0
idna==3.6
numpy==1.26.4
outcome==1.3.0.post0        
packaging==24.0
pandas==2.2.1
pycodestyle==2.11.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1        
pytz==2024.1
requests==2.31.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0     
soupsieve==2.5
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1      
typing_extensions==4.10.0   
tzdata==2024.1
urllib3==1.26.18
webdriver-manager==4.0.1    
wsproto==1.2.0
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.3
djangorestframework==3.15.1
h11==0.14.0
idna==3.6
lxml==5.1.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pycodestyle==2.11.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
webdriver-manager==4.0.1
wsproto==1.2.0
